fixed bugs and added new Scene fixed

+++ BrotatoPrototype/Assets/Scripts/Enemy/EnemyHealth.cs
@@ -30,11 +30,11 @@ public class EnemyHealth : LivingEntity
         base.TakeHit(damage, isCrit);
         if (isCrit)
         {
-            TemporaryMessageManager.Instance.AddMessageOnScreen(damage.ToString() + "!", this.gameObject.transform.position, Color.yellow, 0.5f, 20);
+            TemporaryMessageManager.Instance.AddMessageOnScreen(((int)damage).ToString() + "!", this.gameObject.transform.position, Color.yellow, 0.5f, 20);
         }
         else
         {
-            TemporaryMessageManager.Instance.AddMessageOnScreen(damage.ToString(), this.gameObject.transform.position, Color.white, 0.5f, 20);
+            TemporaryMessageManager.Instance.AddMessageOnScreen(((int)damage).ToString(), this.gameObject.transform.position, Color.white, 0.5f, 20);
         }
         
     }
